Music now fades, updated class data for sfx
Class data now has a new class on it that we will add to for sound effects specific to the class. If there are no specific sounds then we'll probably just play it off of a generic sound data that will be stored elsewhere.

Music now fades similar to RPRT. Still no fades between areas as of yet.