Added QoL deck selector, started end of duel effs
End of duel now has more of a queue system where it waits for each effect to happen before the next does. Right now Slippery solution will softlock the game as it is not completed yet. Also deck selector for movement and support decks added to the start debug menu for easy access to swapping decks for testing even in builds.