Added end of duel space effect functionality
Also updated version number. We can now have spaces like Hall of rest do their effect when you win a duel. Right now the 'popup' for the artwork will be visible twice though once at the end of the duel and once at the end of the turn. Will need to fix that.